Rename ppt package to powerpoint for consistency
- Renamed all source directories from 'ppt' to 'powerpoint'
- Updated import statements across the codebase
- Maintained same file structure and functionality
- Applied consistent naming convention with other packages
