+++ core/src/main/scala/bridges/powerpoint/SlidesDataJsonBridge.scala
@@ -2,9 +2,9 @@ package com.tjclp.xlcr
 package bridges.powerpoint
 
 import bridges.MergeableSymmetricBridge
-import models.ppt.SlidesData
-import parsers.ppt.SlidesDataJsonParser
-import renderers.ppt.SlidesDataJsonRenderer
+import models.powerpoint.SlidesData
+import parsers.powerpoint.SlidesDataJsonParser
+import renderers.powerpoint.SlidesDataJsonRenderer
 import types.MimeType
 import types.MimeType.ApplicationJson
