+++ core/src/main/scala/models/powerpoint/SlideData.scala
@@ -1,5 +1,5 @@
 package com.tjclp.xlcr
-package models.ppt
+package models.powerpoint
 
 import types.Mergeable
