+++ core/src/main/scala/models/powerpoint/SlideElementStyle.scala
@@ -1,5 +1,5 @@
 package com.tjclp.xlcr
-package models.ppt
+package models.powerpoint
 
 import io.circe._
 import io.circe.derivation.{Configuration, ConfiguredDecoder, ConfiguredEncoder}
